excel Export design

+++ app/Http/Controllers/DepHasPositionController.php
@@ -9,20 +9,28 @@ use App\Position;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+
 /** *Here is DepHasPositionController to show,store,insert,update,delete,search data
-     * * @author HZ
-     * @create date 28/08/2020* */
+ * * @author HZ
+ * @create date 28/08/2020 * */
 class DepHasPositionController extends Controller
 {
     /**
      * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
+     *@author HZ
+     * @return DepHasPosition table  data
+     * @create date 29/08/2020
      */
     public function index()
     {
         $dep_has_positions = DepHasPosition::all();
-        return $dep_has_positions;
+        if ($dep_has_positions) {
+            return response()->json($dep_has_positions, 200);
+        } else {
+            return response()->json([
+                "Error:500" => "Internal Server Error!"
+            ], 500);
+        }
     }
 
     /**
@@ -39,12 +47,15 @@ class DepHasPositionController extends Controller
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
+     * @author HZ
      * @return DepHasPositions Model
+     * @create date 29/08/2020
      */
     public function store(Request $request)
     {
         //validation check
-        $validator = Validator::make($request->all(),
+        $validator = Validator::make(
+            $request->all(),
             [
                 'department_id' => 'required',
                 'position_id' => 'required'
@@ -53,11 +64,17 @@ class DepHasPositionController extends Controller
         if ($validator->fails()) { //if validation =false
             return response()->json($validator->errors(), 422);
         } else {
-        $dep_has_positions = new DepHasPosition();
-        $dep_has_positions->department_id = $request->department_id;
-        $dep_has_positions->position_id = $request->position_id;
-        $dep_has_positions->save();
-        return $dep_has_positions;
+            try {
+                $dep_has_positions = new DepHasPosition();
+                $dep_has_positions->department_id = $request->department_id;
+                $dep_has_positions->position_id = $request->position_id;
+                $dep_has_positions->save();
+                return response()->json($dep_has_positions, 200);
+            } catch (\Throwable $th) {
+                return response()->json([
+                    "Error:500" => "Save Unsuccessful!"
+                ], 500);
+            }
         }
     }
 
@@ -69,9 +86,15 @@ class DepHasPositionController extends Controller
      */
     public function show($id)
     {
-        $dep_has_positions = DepHasPosition::whereId($id)->firstOrFail();
+        $dep_has_positions = DepHasPosition::whereId($id)->first();
 
-        return $dep_has_positions;
+        if ($dep_has_positions) {
+            return response()->json($dep_has_positions, 200);
+        } else {
+            return response()->json([
+                "Error: 400" => "Bad Input Request"
+            ], 400);
+        }
     }
 
     /**
@@ -94,11 +117,17 @@ class DepHasPositionController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $dep_has_positions = DepHasPosition::whereId($id)->firstOrFail();
-        $dep_has_positions->department_id = $request->department_id;
-        $dep_has_positions->position_id = $request->position_id;
-        $dep_has_positions->update();
-        return $dep_has_positions;
+        try {
+            $dep_has_positions = DepHasPosition::whereId($id)->firstOrFail();
+            $dep_has_positions->department_id = $request->department_id;
+            $dep_has_positions->position_id = $request->position_id;
+            $dep_has_positions->update();
+            return response()->json($dep_has_positions, 200);
+        } catch (\Throwable $th) {
+            return response()->json([
+                "Error:500" => "Internal Server Error"
+            ], 500);
+        }
     }
 
     /**
@@ -109,39 +138,34 @@ class DepHasPositionController extends Controller
      */
     public function destroy($id)
     {
-        $dep_has_positions = DepHasPosition::whereId($id)->firstOrFail();
-        $dep_has_positions->delete();
-        return $dep_has_positions;
+        $dep_has_positions = DepHasPosition::whereId($id)->first();
+
+        if ($dep_has_positions) {
+            $dep_has_positions->delete();
+            return response()->json([
+                "message" => "Deleted"
+            ], 200);
+        } else {
+            return response()->json([
+                "Error" => "Id not found"
+            ], 400);
+        }
     }
 
     /** *Here is Force Delete  
      * * @author HZ
      * @create date 31/08/2020
      * @param id @return Deleted Message * */
-    public function fdelete($id)
+    /*public function fdelete($id)
     {
 
         try {
             $emp_dep_pos = EmpDepPosition::where('department_id', $id)->firstOrFail();
-            // if ($emp_dep_pos) {
-            //     EmpDepPosition::where('department_id', $id)->forcedelete(); //or
-            //     //$emp_dep_pos->forcedelete();
-
-            // }
-
-            // $emp_has_pos = DepHasPosition::where('department_id', $id)->firstOrFail();
-            // if ($emp_has_pos) {
-            //     //$emp_has_pos->forcedelete();
-            // }
-            // $dep = Department::whereId($id)->firstOrFail();
-            // if ($dep) {
-            //     $dep->forcedelete();
-            // }
             return response()->json([
                 "message" => "Deleted"
             ]);
         } catch (Exception $e) {
             return response($e->getMessage());
         }
-    }
+    }*/
 }
